♻️ Refactor fetching supported week data
