+++ app/models/statistics-store/statistics-supported-store.ts
@@ -12,25 +12,15 @@ import {
   StatisticsStoreFetchWeekOptions,
 } from "./statistics-store"
 import {
-  StatisticsSupportedCreatorModel,
-} from "./statistics-supported-creator"
-import {
-  StatisticsSupportedContentModel,
-} from "./statistics-supported-content"
-import {
-  StatisticsSupportedDayModel,
-} from "./statistics-supported-day"
-import {
   StatisticsSupportedWeekModel,
 } from "./statistics-supported-week"
 
 import { CreatorModel, Creator } from "../creator"
 
-import { logError } from "../../utils/error"
 import {
-  StatisticsSupportedResult,
   StatisticsTopSupportedCreatorsResult,
 } from "../../services/api"
+import { logError } from "../../utils/error"
 
 /**
  * Store for supported statistics
@@ -72,53 +62,7 @@ export const StatisticsSupportedStoreModel = StatisticsStoreModel
       if (opts.shouldSelect) {
         self.selectedWeek = week
       }
-      week.setFetching()
-      try {
-        const result: StatisticsSupportedResult =
-          yield self.env.likeCoAPI.fetchSupportedStatistics(
-            startTs,
-            week.getEndDate().valueOf()
-          )
-        if (result.kind !== "ok") {
-          throw new Error("STATS_FETCH_SUPPORTED_FAILED")
-        }
-
-        week.setWorksCount(result.data.workCount)
-        week.setLikesCount(result.data.likeCount)
-        week.setLikeAmount(result.data.LIKE)
-        week.setCreators(result.data.all.map(({ likee, workCount, LIKE, likeCount }) => {
-          const creator = StatisticsSupportedCreatorModel.create({
-            likerID: likee,
-            likeAmount: LIKE,
-            likesCount: likeCount,
-            worksCount: workCount,
-          }, self.env)
-          creator.setInfo(self.readerStore.createCreatorFromLikerId(likee))
-          return creator
-        }))
-        week.setDays(
-          result.data.daily.map((contents, i) => {
-            const dayID = `${startTs}-${i + 1}`
-            const day = StatisticsSupportedDayModel.create({ dayID }, self.env)
-            if (contents && contents.length > 0) {
-              day.setContents(contents.map(({ sourceURL: url, LIKE, likeCount }) => {
-                const content = StatisticsSupportedContentModel.create({
-                  id: url,
-                  likeAmount: LIKE,
-                  likesCount: likeCount,
-                }, self.env)
-                content.setInfo(self.readerStore.getContentByURL(url))
-                return content
-              }))
-            }
-            return day
-          })
-        )
-      } catch (error) {
-        logError(error.message)
-      } finally {
-        week.setFetched()
-      }
+      yield week.fetchData()
       return week
     }),
     fetchTopSupportedCreators: flow(function * () {
